tests/wpt/tests/html/canvas/element/path-objects/2d.path.roundrect.selfintersect.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.path.roundrect.selfintersect</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.path.roundrect.selfintersect</h1>
<p class="desc"></p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("");
_addTest(function(canvas, ctx) {

  ctx.fillStyle = '#f00';
  ctx.roundRect(0, 0, 100, 50, [0]);
  ctx.strokeStyle = '#0f0';
  ctx.lineWidth = 90;
  ctx.beginPath();
  ctx.roundRect(45, 20, 10, 10, [0]);
  ctx.stroke();
  _assertPixel(canvas, 50,25, 0,255,0,255);

});
</script>
